refactored session data storage and export
